projects/script-starter-py/pyproject.toml
[tool.poetry]
name = "script-starter-py"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["Brian Whippo <brian.whippo@algorand.foundation>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^3.0.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

